requirements.txt
-e git+https://github.com/AIcrowd/aicrowd-cli.git@0903f0c2d95c2938627e0badc74ad2f7e636b02a#egg=aicrowd_cli
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==5.0.2
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
flake8-docstrings==1.5.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.4.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.1.0
packaging==20.0
pip-check-reqs==2.0.3
pip-licenses==1.18.0
pkginfo==1.5.0.1
pluggy==0.13.1
PTable==0.9.2
py==1.8.1
pycodestyle==2.5.0
pycryptodome==3.9.4
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
pyperclip==1.7.0
pytest==3.10.1
pytest-cov==2.8.1
pytest-pythonpath==0.7.3
pytz==2019.3
PyYAML==5.3
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
shellescape==3.4.1
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
ssh-agent-setup==1.0.1
toml==0.10.0
tox==3.14.3
tqdm==4.41.1
twine==1.15.0
urllib3==1.25.7
virtualenv==16.7.9
webencodings==0.5.1
wget==3.2
zipp==0.6.0
